requirements.txt
Flask>=2.3.3
Flask-CORS>=4.0.0
openai>=1.3.8
gunicorn>=21.2.0
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.7
python-docx>=1.1.2
pypandoc>=1.12